upgrade s3

+++ lib/carrierwave/encrypter_decrypter/openssl/aes.rb
@@ -12,18 +12,25 @@ module Openssl
         model.key = key
         model.save!
 
-        original_file_path = File.expand_path(obj.store_path, obj.root)
-        encrypted_file_path = File.expand_path(obj.store_path, obj.root) + ".enc"
-        buf = ""
-        File.open(encrypted_file_path, "wb") do |outf|
-          File.open(model.send(mounted_as).path, "rb") do |inf|
-            while inf.read(4096, buf)
-              outf << cipher.update(buf)
+        if Carrierwave::EncrypterDecrypter::StorageHelper.fog_storage?(obj)
+          content = obj.file.read
+          encrypted_data = cipher.update(content) + cipher.final
+          Carrierwave::EncrypterDecrypter::StorageHelper.fog_write(obj, obj.store_path + '.enc', encrypted_data)
+          obj.file.delete
+        else
+          original_file_path = File.expand_path(obj.store_path, obj.root)
+          encrypted_file_path = File.expand_path(obj.store_path, obj.root) + ".enc"
+          buf = ""
+          File.open(encrypted_file_path, "wb") do |outf|
+            File.open(model.send(mounted_as).path, "rb") do |inf|
+              while inf.read(4096, buf)
+                outf << cipher.update(buf)
+              end
+              outf << cipher.final
             end
-            outf << cipher.final
           end
+          File.unlink(model.send(mounted_as).path)
         end
-        File.unlink(model.send(mounted_as).path)
       rescue Exception => e
         puts "****************************#{e.message}"
         puts "****************************#{e.backtrace.inspect}"
@@ -38,17 +45,25 @@ module Openssl
         cipher.decrypt
         cipher.iv = model.iv
         cipher.key = model.key
-        buf = ""
 
-        original_file_path =  obj.send(mounted_as).root + obj.send(mounted_as).url
-        encrypted_file_path =  obj.send(mounted_as).root + obj.send(mounted_as).url  + ".enc"
+        uploader = obj.send(mounted_as)
 
-        File.open(original_file_path, "wb") do |outf|
-          File.open(encrypted_file_path, "rb") do |inf|
-            while inf.read(4096, buf)
-              outf << cipher.update(buf)
+        if Carrierwave::EncrypterDecrypter::StorageHelper.fog_storage?(uploader)
+          encrypted_data = Carrierwave::EncrypterDecrypter::StorageHelper.fog_read(uploader, uploader.store_path + '.enc')
+          decrypted_data = cipher.update(encrypted_data) + cipher.final
+          Carrierwave::EncrypterDecrypter::StorageHelper.fog_write(uploader, uploader.store_path, decrypted_data)
+        else
+          buf = ""
+          original_file_path = obj.send(mounted_as).root + obj.send(mounted_as).url
+          encrypted_file_path = obj.send(mounted_as).root + obj.send(mounted_as).url + ".enc"
+
+          File.open(original_file_path, "wb") do |outf|
+            File.open(encrypted_file_path, "rb") do |inf|
+              while inf.read(4096, buf)
+                outf << cipher.update(buf)
+              end
+              outf << cipher.final
             end
-            outf << cipher.final
           end
         end
       rescue Exception => e
